initial multi-setup support for TFTchar scripts
git-svn-id: svn://svn.imager.umro/masdax/trunk/scripts/scriptmeas@545 92890ac3-fb6f-4dbc-9a08-888296f68159

+++ multim_2636A_ctrlTFT.sh
@@ -3,8 +3,15 @@
 # $HeadURL$
 # $Id$
 
-DATACTRLFILE='commtemp/2636_datactrl'
-SCPIFILE='commtemp/2636_command.scpi'
+#DATACTRLFILE='commtemp/2636_datactrl'
+#SCPIFILE='commtemp/2636_command.scpi'
+
+# source the settings file
+source "$1"
+if [ "$MAINSMU" == "" ]; then
+ echo "Important variables not set. Did you specify the settings file?"
+ exit 5
+fi
 
 rm "$SCPIFILE"
